Auto-determine subproject names
This removes the implicit assumption that the name with which a
dependency is declared matches the name of the dependency's project as
used in their source code (including casing).
Instead, the project name is now parsed from the top-level
CMakeLists.txt file and is therefore decoupled from the name given while
declaring the dependency.

Fixes #2